refactoring PriceLevelIcons
